pyproject.toml
[tool.poetry]
name = "inventify"
version = "0.1.0"
description = ""
authors = ["Lisa <bakosh21345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
django-simple-history = "^3.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
pillow = "^9.5.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
loguru = "^0.7.2"
django-filter = "^23.4"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-results = "^2.5.1"
django-json-widget = "^2.0.1"
flower = "^2.0.1"
python-dotenv = "^1.0.1"
django-redis = "^5.4.0"
django-eav2 = "^1.6.1"
gevent = "^24.2.1"
django-silk = "^5.1.0"
djangoql = "0.17.1"
pydantic = "^2.9.2"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
